flask-server/requirements.txt
alembic==1.9.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
id==0.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==22.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
